chore: add back to project button
